requirements.txt
numpy==1.26.1
pandas==2.1.2
Pillow==10.1.0
torch==1.13.0
